Removed the 'Software Templates' (https://janus-idp.io/gpts) page from the Janus website as part of our sunsetting efforts. Pointed any 'Software Templates' link to (https://github.com/janus-idp/software-templates) as a historical reference.
I also included a new tab icon next to the 'Software Templates' link to provide context to the user that they will be directed to a new tab. This also increases consistency as to which links allow the user to stay in the site and which links direct the user to a new tab.