chore: fix following bump to Husky v9.1.1 (#335)
Not sure why we are building the website from PRs in the
backstage-plugins repo, but PR checks in the
janus-idp/backstage-plugins repo are failing due to an error
about dirty workspaces. See [1].

This has been narrowed down to the bump of Husky to v9.1.1 [2],
which, according to the release notes [3], makes Husky automatically
remove some deprecated statements.

[1] https://github.com/janus-idp/backstage-plugins/actions/workflows/pr-website.yaml
[2] https://github.com/janus-idp/janus-idp.github.io/pull/330
[3] https://github.com/typicode/husky/releases/tag/v9.1.1